--- src/hooks/useMedia.js
@@ -1,22 +1,0 @@
-import { useEffect, useState } from "react";
-
-const useMediaQuery = (query) => {
-  const [matches, setMatches] = useState(false);
-
-  useEffect(() => {
-    const mediaQuery = window.matchMedia(query);
-    setMatches(mediaQuery.matches);
-
-    const handleMediaQuery = (e) => {
-      setMatches(e.matches);
-    };
-
-    mediaQuery.addEventListener("change", handleMediaQuery);
-
-    return () => mediaQuery.removeEventListener("change", handleMediaQuery);
-  }, [query]);
-
-  return matches;
-};
-
-export default useMediaQuery;
